requirements.txt
addict
appdirs
attrdict
attrs
begins
black
click
coverage
colorama
cryptography
cycler
cython
dask
decorator
einops
falcon
fastapi[all]==0.85.2
h5py
hdbscan
httpx
hug
imageio==2.9.0
ipywidgets==7.7.2
jupyterlab
kornia
loguru
mako
matplotlib==3.3.1
medicaltorch
monai
morphsnakes
mrcfile
napari[all]
networkx
nibabel
nose
numba
numpy
numpydoc
pandas==1.5.3
parse
pillow
pip
pluggy
ptyprocess
pydantic
pygments
pyopengl
pyqtgraph==0.11.1
pysocks
pytest
python-snappy
pytools
pytz
pywavelets
pyyaml
pyzmq
PyQt5==5.14.2
qtawesome
qtconsole
qtpy
requests
ruamel.yaml
scikit-image==0.19.2
scikit-learn
scipy
seaborn
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
setuptools==65.6.3
simpleitk
snowballstemmer
tensorly
tifffile
torch
torchio
torchmetrics
tqdm
traitlets
transformers
trio
typing_extensions
umap-learn
unet
urllib3
vispy
volume-segmantics
wget
wheel
zict
zstd
